+++ fsleyes/gl/gl21/glvector_funcs.py
@@ -55,6 +55,7 @@ def updateShaderState(self, useSpline=False):
     opts              = self.opts
     shader            = self.shader
     imageShape        = self.vectorImage.shape[:3]
+    normaliseColour   = opts.normaliseColour
     modLow,  modHigh  = self.getModulateRange()
     clipLow, clipHigh = self.getClippingRange()
 
@@ -63,8 +64,9 @@ def updateShaderState(self, useSpline=False):
     if opts.clipImage     is None: clipShape = [1, 1, 1]
     else:                          clipShape = opts.clipImage.shape[:3]
 
-    modMode     = {'brightness' : 0,
-                   'alpha'      : 1}[opts.modulateMode]
+    if   opts.modulateMode == 'brightness': modMode = 0
+    elif opts.modulateMode == 'alpha':      modMode = 1
+    else:                                   modMode = -1
 
     clipXform   = self.getAuxTextureXform('clip')
     colourXform = self.getAuxTextureXform('colour')
@@ -116,6 +118,7 @@ def updateShaderState(self, useSpline=False):
         changed |= shader.set('xColour',         colours[0])
         changed |= shader.set('yColour',         colours[1])
         changed |= shader.set('zColour',         colours[2])
+        changed |= shader.set('normaliseColour', normaliseColour)
         changed |= shader.set('colourXform',     colourXform)
         changed |= shader.set('voxValXform',     voxValXform)
         changed |= shader.set('imageShape',      imageShape)