+++ fsleyes/gl/glrgbvector.py
@@ -81,34 +81,19 @@ class GLRGBVector(glvector.GLVector):
 
         def prefilter(data):
 
-            # Vector images stored as RGB24 data
-            # type are assumed to map from [0, 255]
-            # to [-1, 1], so cannot be normalised
+            # Vector images stored as RGB24 data type
+            # contain values from [0, 255] so cannot
+            # be made absolute
             if vecImage.nvals > 1:
                 return data
 
-            # make absolute, and scale to unit
-            # length if required. We must make
-            # the data absolute, otherwise we
-            # cannot perform interpolation on
-            # the texture when displaying it.
-            data = np.abs(data)
-            if self.opts.unitLength:
-                with np.errstate(invalid='ignore'):
-                    x            = data[0, ...]
-                    y            = data[1, ...]
-                    z            = data[2, ...]
-                    lens         = np.sqrt(x ** 2 + y ** 2 + z ** 2)
-                    data[0, ...] = x / lens
-                    data[1, ...] = y / lens
-                    data[2, ...] = z / lens
-            return data
+            # We must make the vector values absolute,
+            # otherwise we cannot perform interpolation
+            # on the texture when displaying it.
+            return np.abs(data)
 
         def prefilterRange(dmin, dmax):
-            if self.opts.unitLength:
-                return 0, 1
-            else:
-                return max((0, dmin)), max((abs(dmin), abs(dmax)))
+            return max((0, dmin)), max((abs(dmin), abs(dmax)))
 
         glvector.GLVector.__init__(self,
                                    image,
@@ -124,9 +109,9 @@ class GLRGBVector(glvector.GLVector):
         self.opts.addListener('interpolation',
                               self.name,
                               self.__interpChanged)
-        self.opts.addListener('unitLength',
+        self.opts.addListener('normaliseColour',
                               self.name,
-                              self.__unitLengthChanged)
+                              self.asyncUpdateShaderState)
 
 
     def destroy(self):
@@ -135,8 +120,8 @@ class GLRGBVector(glvector.GLVector):
         instance, calls the OpenGL version-specific ``destroy``
         function, and calls the :meth:`.GLVector.destroy` method.
         """
-        self.opts.removeListener('interpolation', self.name)
-        self.opts.removeListener('unitLength',    self.name)
+        self.opts.removeListener('interpolation',   self.name)
+        self.opts.removeListener('normaliseColour', self.name)
         fslgl.glrgbvector_funcs.destroy(self)
         glvector.GLVector.destroy(self)
 
@@ -182,13 +167,6 @@ class GLRGBVector(glvector.GLVector):
         self.asyncUpdateShaderState(alwaysNotify=True)
 
 
-    def __unitLengthChanged(self, *a):
-        """Called when :attr:`.RGBVectorOpts.unitLength` changes. Refreshes
-        the texture data.
-        """
-        self.imageTexture.refresh()
-
-
     def compileShaders(self):
         """Overrides :meth:`.GLVector.compileShaders`. Calls the OpenGL
         version-specific ``compileShaders`` function.
@@ -197,7 +175,7 @@ class GLRGBVector(glvector.GLVector):
 
 
     def updateShaderState(self):
-        """Overrides :meth:`.GLVector.compileShaders`. Calls the OpenGL
+        """Overrides :meth:`.GLVector.updateShaderState`. Calls the OpenGL
         version-specific ``updateShaderState`` function.
         """
         return fslgl.glrgbvector_funcs.updateShaderState(self)