+++ fsleyes/tests/overlays/test_overlay_rgbvector.py
@@ -33,17 +33,17 @@ dti/dti_V1 -ot rgbvector -in spline -b 25 -c 25
 
 {{asrgb('dti/dti_V1')}} -ot rgbvector
 
-# unit length scaling
-{{mul('dti/dti_V1', 0.5)}} -ot rgbvector -u
+# Colour normalisation
+{{mul('dti/dti_V1', 0.5)}} -ot rgbvector -nr
 {{mul('dti/dti_V1', 0.5)}} -ot rgbvector
-{{mul('dti/dti_V1', 2.0)}} -ot rgbvector -u
+{{mul('dti/dti_V1', 2.0)}} -ot rgbvector -nr
 {{mul('dti/dti_V1', 2.0)}} -ot rgbvector
 
 # Images with intent code 2003 should be
 # automaticallt shown as RGB vectors, so
 # there shouldn't be any need to specify
 # The overlay type at the command-line.
-{{mul('dti/dti_V1', 3.0)}} -in spline -u
+{{mul('dti/dti_V1', 3.0)}} -in spline -nr
 """
 
 extras = {
@@ -60,5 +60,5 @@ def test_overlay_rgbvector_nofloattextures():
     texdata.canUseFloatTextures.invalidate()
     with mock.patch('fsleyes.gl.textures.data.canUseFloatTextures',
                     return_value=(False, None, None)):
-        run_cli_tests('test_overlay_rgbvector_nofloattextures',
+        run_cli_tests('test_overlay_rgbvector',
                       cli_tests, extras=extras)